Expose token errors through AI gateway client
These failures arise exclusively while acquiring a Convex service token, so a separate top-level `AiGatewayError` namespace obscured which API owned them and split their definitions from the classification logic.

Keep the tagged errors alongside the service-token service and re-export them through `AiGatewayClient`. This makes the error channel discoverable from the client that produces it while retaining public access to the classes and union type.
