+++ packages/server/src/internal/AiGatewayServiceToken.ts
@@ -2,11 +2,39 @@ import { getServiceToken as getConvexServiceToken } from "convex/server";
 import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
-import {
-  AiGatewayDisabled,
-  type AiGatewayError,
-  AiGatewayUnavailable,
-} from "../AiGatewayError";
+import * as Schema from "effect/Schema";
+
+/**
+ * The current team is on the free plan or has the Convex AI gateway disabled.
+ * Upgrade to a paid plan, or contact Convex support if this looks wrong.
+ */
+export class AiGatewayDisabled extends Schema.TaggedError<AiGatewayDisabled>()(
+  "AiGatewayDisabled",
+  {},
+) {
+  override get message(): string {
+    return "The Convex AI gateway is disabled. Your team may be on the free plan, or the gateway may have been disabled for your team. Upgrade to a paid plan, or email support@convex.dev if this looks wrong.";
+  }
+}
+
+/**
+ * The current deployment is local or self-hosted and cannot use the Convex AI
+ * gateway. Call the model provider directly with your own API key instead.
+ */
+export class AiGatewayUnavailable extends Schema.TaggedError<AiGatewayUnavailable>()(
+  "AiGatewayUnavailable",
+  {},
+) {
+  override get message(): string {
+    return "The Convex AI gateway is unavailable. This action is running on a local or self-hosted deployment, which cannot use the gateway. Call the model provider directly with your own API key stored in a Convex environment variable.";
+  }
+}
+
+/**
+ * The possible failures that can occur while obtaining a Convex AI gateway
+ * service token.
+ */
+export type AiGatewayError = AiGatewayDisabled | AiGatewayUnavailable;
 
 type GetServiceToken = typeof getConvexServiceToken;
 type AiGatewayErrorCode = "AiGatewayDisabled" | "AiGatewayUnavailable";
