+++ packages/server/test/AiGatewayServiceToken.test.ts
@@ -1,4 +1,4 @@
-import { AiGatewayError } from "@confect/server";
+import { AiGatewayClient } from "@confect/server";
 import { assert, describe, it } from "@effect/vitest";
 import * as Effect from "effect/Effect";
 import * as Exit from "effect/Exit";
@@ -17,7 +17,7 @@ describe("AiGatewayServiceToken", () => {
         new Error(CONVEX_AI_GATEWAY_DISABLED_MESSAGE),
       );
 
-      assert.instanceOf(error, AiGatewayError.AiGatewayDisabled);
+      assert.instanceOf(error, AiGatewayClient.AiGatewayDisabled);
       assert.strictEqual(
         error.message,
         "The Convex AI gateway is disabled. Your team may be on the free plan, or the gateway may have been disabled for your team. Upgrade to a paid plan, or email support@convex.dev if this looks wrong.",
@@ -35,7 +35,7 @@ describe("AiGatewayServiceToken", () => {
         ),
       );
 
-      assert.instanceOf(error, AiGatewayError.AiGatewayDisabled);
+      assert.instanceOf(error, AiGatewayClient.AiGatewayDisabled);
     }),
   );
 
@@ -45,7 +45,7 @@ describe("AiGatewayServiceToken", () => {
         new Error(CONVEX_AI_GATEWAY_UNAVAILABLE_MESSAGE),
       );
 
-      assert.instanceOf(error, AiGatewayError.AiGatewayUnavailable);
+      assert.instanceOf(error, AiGatewayClient.AiGatewayUnavailable);
       assert.strictEqual(
         error.message,
         "The Convex AI gateway is unavailable. This action is running on a local or self-hosted deployment, which cannot use the gateway. Call the model provider directly with your own API key stored in a Convex environment variable.",
@@ -63,7 +63,7 @@ describe("AiGatewayServiceToken", () => {
         ),
       );
 
-      assert.instanceOf(error, AiGatewayError.AiGatewayUnavailable);
+      assert.instanceOf(error, AiGatewayClient.AiGatewayUnavailable);
     }),
   );
